Options --camera & --input, as well as everyFrame functionality.
--camera allows specification of OpenCV live camera feed identifier
numbers (default: 0).

--input allows specification of input MP4 files.

If --input is provided it overrides --camera.

If --input is provided then 'everyFrame' is enabled, which causes
synchronising events to be expected and generated as every frame is
processed, to ensure that no frame is dropped. Live camera feed disables
'everyFrame', meaning that there is no synchronisation between input
frames and output frames so that frames can be dropped. This is
important because if you do not pull every frame from OpenCV in
'real-time' then they start to queue up and you start backlogging and
processing frames from a few seconds ago or worse. (Remember, the
'queue' that connects the first stage of the pipeline is really more
like an 'mbox' that doesn't queue items, instead it overwrites whatever
is waiting, on the assumption that it has become stale if it is still
sitting in the 'queue').